NEW: `query_transaction_tree` function returns message and transaction tree produced by specified message. FIX: `AbiData.key` type changed to u32. FIX: attempt to use `orderBy` instead of `order` in `query_collection` will raise error.

+++ ton_client/src/abi/types.rs
@@ -76,7 +76,7 @@ pub struct AbiEvent {
 
 #[derive(Serialize, Deserialize, Clone, Debug, ApiType, Default)]
 pub struct AbiData {
-    pub key: u64,
+    pub key: u32,
     pub name: String,
     #[serde(rename = "type")]
     pub param_type: String,
